Use BigInt for int64/uint64 in readers and writers
- SpecReader.readInt64()/readUint64() now return BigInt
- MsgPackReader/JsonReader/GronReader updated to return BigInt
- MsgPackWriter/JsonWriter.writeInt64()/writeUint64() now accept BigInt
- Ensures cross-platform compatibility (web platform int is 53-bit)

+++ lib/src/spec_reader.dart
@@ -11,9 +11,9 @@ abstract class SpecReader {
   String readString();
   bool readBool();
   int readInt32();
-  int readInt64();
+  BigInt readInt64();
   int readUint32();
-  int readUint64();
+  BigInt readUint64();
   double readFloat32();
   double readFloat64();
   void readNull();
